chore: add view table definition in table detail drawer
