+++ frontend/src/locales/zh-CN.json
@@ -1749,7 +1749,8 @@
     "partition-tables": "分区表",
     "external-server-name": "外部服务器名称",
     "external-database-name": "外部数据库名称",
-    "foreign-table-detail": "外部表详情"
+    "foreign-table-detail": "外部表详情",
+    "view-definition": "查看定义"
   },
   "repository": {
     "branch-observe-file-change": "Bytebase 跟踪文件变更的分支。",
